Explore new OpenAlex data and refactor
