+++ discovery_child_development/utils/openai_utils.py
@@ -1,4 +1,7 @@
 import openai
 import os
+from dotenv import load_dotenv
+
+load_dotenv()
 
 openai.api_key = os.environ["OPENAI_API_KEY"]
